frontend/pyproject.toml
[project]
name = "traffic-incident-ui"
version = "0.1.0"
description = "FastHTML Dashboard for Traffic Incident Monitoring"
requires-python = ">=3.10"
dependencies = [
    "python-fasthtml>=0.1.0",
    "httpx>=0.27.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.0.0"
]

[tool.uv]
dev-dependencies = ["pytest"]
